Added dark mode feature to portfolio
Icons files updated so that there are icons for light mode and dark mode. HTML updated to add these new icons. A new icon added for toggling between light and dark mode. Functionality created in the .js file so that clicking on the icon actually does something and the icon also changes with the colour of everything else. Finally, the static colours were changed to variables so that they adjusted according to the theme of the site.